feat: Add JWT authentication and token-based access control
