+++ .env.example
@@ -968,3 +968,9 @@ BUNDLE_EXECUTOR_ADDRESS=          # Address of the bundle executor contract
 DESK_EXCHANGE_PRIVATE_KEY=                  # Required for trading and cancelling orders
 DESK_EXCHANGE_NETWORK=                      # "mainnet" or "testnet
 
+# JWT
+JWT_ENABLED=
+JWT_SECRET_KEY=
+JWT_EXPIRED=
+JWT_USERNAME=
+JWT_PASSWORD=
