+++ packages/client-direct/src/api.ts
@@ -8,47 +8,16 @@ import {
     type AgentRuntime,
     elizaLogger,
     getEnvVariable,
-    type UUID,
-    validateCharacterConfig,
     ServiceType,
     type Character,
+    settings,
 } from "@elizaos/core";
 
 import type { TeeLogQuery, TeeLogService } from "@elizaos/plugin-tee-log";
 import { REST, Routes } from "discord.js";
 import type { DirectClient } from ".";
-import { validateUuid } from "@elizaos/core";
-
-interface UUIDParams {
-    agentId: UUID;
-    roomId?: UUID;
-}
-
-function validateUUIDParams(
-    params: { agentId: string; roomId?: string },
-    res: express.Response
-): UUIDParams | null {
-    const agentId = validateUuid(params.agentId);
-    if (!agentId) {
-        res.status(400).json({
-            error: "Invalid AgentId format. Expected to be a UUID: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
-        });
-        return null;
-    }
-
-    if (params.roomId) {
-        const roomId = validateUuid(params.roomId);
-        if (!roomId) {
-            res.status(400).json({
-                error: "Invalid RoomId format. Expected to be a UUID: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
-            });
-            return null;
-        }
-        return { agentId, roomId };
-    }
-
-    return { agentId };
-}
+import { validateUUIDParams } from ".";
+import { md5, signToken } from "./auth";
 
 export function createApiRouter(
     agents: Map<string, AgentRuntime>,
@@ -64,14 +33,6 @@ export function createApiRouter(
             limit: getEnvVariable("EXPRESS_MAX_PAYLOAD") || "100kb",
         })
     );
-
-    router.get("/", (req, res) => {
-        res.send("Welcome, this is the REST API!");
-    });
-
-    router.get("/hello", (req, res) => {
-        res.json({ message: "Hello World!" });
-    });
 
     router.get("/agents", (req, res) => {
         const agentsList = Array.from(agents.values()).map((agent) => ({
@@ -113,102 +74,6 @@ export function createApiRouter(
         res.json({
             id: agent.agentId,
             character: agent.character,
-        });
-    });
-
-    router.delete("/agents/:agentId", async (req, res) => {
-        const { agentId } = validateUUIDParams(req.params, res) ?? {
-            agentId: null,
-        };
-        if (!agentId) return;
-
-        const agent: AgentRuntime = agents.get(agentId);
-
-        if (agent) {
-            agent.stop();
-            directClient.unregisterAgent(agent);
-            res.status(204).json({ success: true });
-        } else {
-            res.status(404).json({ error: "Agent not found" });
-        }
-    });
-
-    router.post("/agents/:agentId/set", async (req, res) => {
-        const { agentId } = validateUUIDParams(req.params, res) ?? {
-            agentId: null,
-        };
-        if (!agentId) return;
-
-        let agent: AgentRuntime = agents.get(agentId);
-
-        // update character
-        if (agent) {
-            // stop agent
-            agent.stop();
-            directClient.unregisterAgent(agent);
-            // if it has a different name, the agentId will change
-        }
-
-        // stores the json data before it is modified with added data
-        const characterJson = { ...req.body };
-
-        // load character from body
-        const character = req.body;
-        try {
-            validateCharacterConfig(character);
-        } catch (e) {
-            elizaLogger.error(`Error parsing character: ${e}`);
-            res.status(400).json({
-                success: false,
-                message: e.message,
-            });
-            return;
-        }
-
-        // start it up (and register it)
-        try {
-            agent = await directClient.startAgent(character);
-            elizaLogger.log(`${character.name} started`);
-        } catch (e) {
-            elizaLogger.error(`Error starting agent: ${e}`);
-            res.status(500).json({
-                success: false,
-                message: e.message,
-            });
-            return;
-        }
-
-        if (process.env.USE_CHARACTER_STORAGE === "true") {
-            try {
-                const filename = `${agent.agentId}.json`;
-                const uploadDir = path.join(
-                    process.cwd(),
-                    "data",
-                    "characters"
-                );
-                const filepath = path.join(uploadDir, filename);
-                await fs.promises.mkdir(uploadDir, { recursive: true });
-                await fs.promises.writeFile(
-                    filepath,
-                    JSON.stringify(
-                        { ...characterJson, id: agent.agentId },
-                        null,
-                        2
-                    )
-                );
-                elizaLogger.info(
-                    `Character stored successfully at ${filepath}`
-                );
-            } catch (error) {
-                elizaLogger.error(
-                    `Failed to store character: ${error.message}`
-                );
-            }
-        }
-
-        res.json({
-            id: character.id,
-            character: character,
         });
     });
 
@@ -404,53 +269,14 @@ export function createApiRouter(
         }
     );
 
-    router.post("/agent/start", async (req, res) => {
-        const { characterPath, characterJson } = req.body;
-        console.log("characterPath:", characterPath);
-        console.log("characterJson:", characterJson);
-        try {
-            let character: Character;
-            if (characterJson) {
-                character = await directClient.jsonToCharacter(
-                    characterPath,
-                    characterJson
-                );
-            } else if (characterPath) {
-                character =
-                    await directClient.loadCharacterTryPath(characterPath);
-            } else {
-                throw new Error("No character path or JSON provided");
-            }
-            await directClient.startAgent(character);
-            elizaLogger.log(`${character.name} started`);
-
-            res.json({
-                id: character.id,
-                character: character,
-            });
-        } catch (e) {
-            elizaLogger.error(`Error parsing character: ${e}`);
-            res.status(400).json({
-                error: e.message,
-            });
-            return;
-        }
-    });
-
-    router.post("/agents/:agentId/stop", async (req, res) => {
-        const agentId = req.params.agentId;
-        console.log("agentId", agentId);
-        const agent: AgentRuntime = agents.get(agentId);
-
-        // update character
-        if (agent) {
-            // stop agent
-            agent.stop();
-            directClient.unregisterAgent(agent);
-            // if it has a different name, the agentId will change
-            res.json({ success: true });
+    router.post("/auth/login", async (req, res) => {
+        const { username, password } = req.body;
+        const valid = username === settings.JWT_USERNAME && password === md5(settings.JWT_PASSWORD);
+        if (valid) {
+            const token = signToken({ username });
+            res.json({ success: true, token: token });
         } else {
-            res.status(404).json({ error: "Agent not found" });
+            res.status(401).json({ error: "Invalid username or password" });
         }
     });
