+++ packages/client-direct/src/manage-api.ts
@@ -9,18 +9,16 @@ import {
     type AgentRuntime,
     elizaLogger,
     getEnvVariable,
-    type UUID,
     validateCharacterConfig,
-    ServiceType,
+    type UUID,
     type Character,
     type PaginationParams,
     AccountStatus,
     stringToUuid
 } from "@elizaos/core";
 
-import type { TeeLogQuery, TeeLogService } from "@elizaos/plugin-tee-log";
-import { REST, Routes } from "discord.js";
 import type { DirectClient } from ".";
+import { validateUUIDParams } from ".";
 
 export function createManageApiRouter(
     agents: Map<string, AgentRuntime>,
@@ -36,6 +34,14 @@ export function createManageApiRouter(
             limit: getEnvVariable("EXPRESS_MAX_PAYLOAD") || "100kb",
         })
     );
+
+    const changeAccountStatus = async (accountId: UUID, status: AccountStatus) => {
+        const account = await directClient.db.getAccountById(accountId);
+        if(account) {
+            account.status = status;
+            await directClient.db.updateAccount(account);
+        }
+    }
 
     router.get("/accounts", async (req, res) => {
         const params: PaginationParams = {
@@ -99,6 +105,153 @@ export function createManageApiRouter(
                 error: e.message,
             });
             return;
+        }
+    });
+
+    router.delete("/agents/:agentId", async (req, res) => {
+        const { agentId } = validateUUIDParams(req.params, res) ?? {
+            agentId: null,
+        };
+        if (!agentId) return;
+
+        const agent: AgentRuntime = agents.get(agentId);
+
+        if (agent) {
+            agent.stop();
+            directClient.unregisterAgent(agent);
+            await changeAccountStatus(agentId, AccountStatus.DISABLED);
+            res.status(204).json({ success: true });
+        } else {
+            res.status(404).json({ error: "Agent not found" });
+        }
+    });
+
+    router.post("/agents/:agentId/set", async (req, res) => {
+        const { agentId } = validateUUIDParams(req.params, res) ?? {
+            agentId: null,
+        };
+        if (!agentId) return;
+
+        let agent: AgentRuntime = agents.get(agentId);
+
+        // update character
+        if (agent) {
+            // stop agent
+            agent.stop();
+            directClient.unregisterAgent(agent);
+            // if it has a different name, the agentId will change
+        }
+
+        // stores the json data before it is modified with added data
+        const characterJson = { ...req.body };
+
+        // load character from body
+        const character = req.body;
+        try {
+            validateCharacterConfig(character);
+        } catch (e) {
+            elizaLogger.error(`Error parsing character: ${e}`);
+            res.status(400).json({
+                success: false,
+                message: e.message,
+            });
+            return;
+        }
+
+        // start it up (and register it)
+        try {
+            agent = await directClient.startAgent(character);
+            elizaLogger.log(`${character.name} started`);
+        } catch (e) {
+            elizaLogger.error(`Error starting agent: ${e}`);
+            res.status(500).json({
+                success: false,
+                message: e.message,
+            });
+            return;
+        }
+
+        if (process.env.USE_CHARACTER_STORAGE === "true") {
+            try {
+                const filename = `${agent.agentId}.json`;
+                const uploadDir = path.join(
+                    process.cwd(),
+                    "data",
+                    "characters"
+                );
+                const filepath = path.join(uploadDir, filename);
+                await fs.promises.mkdir(uploadDir, { recursive: true });
+                await fs.promises.writeFile(
+                    filepath,
+                    JSON.stringify(
+                        { ...characterJson, id: agent.agentId },
+                        null,
+                        2
+                    )
+                );
+                elizaLogger.info(
+                    `Character stored successfully at ${filepath}`
+                );
+            } catch (error) {
+                elizaLogger.error(
+                    `Failed to store character: ${error.message}`
+                );
+            }
+        }
+
+        res.json({
+            id: character.id,
+            character: character,
+        });
+    });
+
+    router.post("/agent/start", async (req, res) => {
+        const { characterPath, characterJson } = req.body;
+        console.log("characterPath:", characterPath);
+        console.log("characterJson:", characterJson);
+        try {
+            let character: Character;
+            if (characterJson) {
+                character = await directClient.jsonToCharacter(
+                    characterPath,
+                    characterJson
+                );
+            } else if (characterPath) {
+                character =
+                    await directClient.loadCharacterTryPath(characterPath);
+            } else {
+                throw new Error("No character path or JSON provided");
+            }
+            await directClient.startAgent(character);
+            elizaLogger.log(`${character.name} started`);
+
+            res.json({
+                id: character.id,
+                character: character,
+            });
+        } catch (e) {
+            elizaLogger.error(`Error parsing character: ${e}`);
+            res.status(400).json({
+                error: e.message,
+            });
+            return;
+        }
+    });
+
+    router.post("/agents/:agentId/stop", async (req, res) => {
+        const agentId = req.params.agentId as UUID;
+        console.log("agentId", agentId);
+        const agent: AgentRuntime = agents.get(agentId);
+
+        // update character
+        if (agent) {
+            // stop agent
+            agent.stop();
+            await changeAccountStatus(agentId, AccountStatus.PAUSED);
+            // if it has a different name, the agentId will change
+            res.json({ success: true });
+        } else {
+            res.status(404).json({ error: "Agent not found" });
         }
     });
